refactor(Company) : Integrate DTOs in the serviceLayer .

+++ Backend/src/main/java/tn/ghassen/Inventory/service/impl/CompanyServiceImpl.java
@@ -2,7 +2,11 @@ package tn.ghassen.inventory.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import tn.ghassen.inventory.dto.company.CompanyCreateDTO;
+import tn.ghassen.inventory.dto.company.CompanyResponseDTO;
+import tn.ghassen.inventory.dto.company.CompanyUpdateDTO;
 import tn.ghassen.inventory.entity.Company;
+import tn.ghassen.inventory.mapper.CompanyMapper;
 import tn.ghassen.inventory.repository.CompanyRepository;
 import tn.ghassen.inventory.service.CompanyService;
 
@@ -13,40 +17,50 @@ import java.util.List;
 class CompanyServiceImpl implements CompanyService {
 
     private final CompanyRepository companyRepository;
+    private final CompanyMapper companyMapper;
 
     @Override
-    public Company createCompany(Company company) {
-        return companyRepository.save(company);
+    public CompanyResponseDTO createCompany(CompanyCreateDTO dto) {
+            Company company = companyMapper.toEntity(dto);
+
+            Company savedCompany = companyRepository.save(company);
+        return companyMapper.toResponseDTO(savedCompany);
     }
 
     @Override
-    public Company getCompanyById(Long id) {
-        return companyRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Company not found"));
+    public CompanyResponseDTO getCompanyById(Long id) {
+        Company company = companyRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Company not found with id "+ id));
+
+        return companyMapper.toResponseDTO(company);
     }
 
     @Override
-    public List<Company> getAllCompanies() {
-        return companyRepository.findAll();
+    public List<CompanyResponseDTO> getAllCompanies() {
+
+        return companyRepository.findAll()
+                .stream()
+                .map(companyMapper::toResponseDTO)
+                .toList();
     }
 
     @Override
-    public Company updateCompany(Long id, Company company) {
+    public CompanyResponseDTO updateCompany(Long id, CompanyUpdateDTO dto) {
 
-        Company existing = getCompanyById(id);
+        Company existing = companyRepository.findById(id)
+                        .orElseThrow(()-> new RuntimeException("Company not found with id "+id));
+        companyMapper.updateEntity(existing, dto);
+        Company updatedCompany = companyRepository.save(existing);
 
-        existing.setName(company.getName());
-        existing.setEmail(company.getEmail());
-        existing.setPhone(company.getPhone());
-        existing.setAddress(company.getAddress());
-        existing.setTaxNumber(company.getTaxNumber());
-        existing.setLogo(company.getLogo());
 
-        return companyRepository.save(existing);
+        return companyMapper.toResponseDTO(updatedCompany);
     }
 
     @Override
     public void deleteCompany(Long id) {
+        if (!companyRepository.existsById(id)) {
+            throw new RuntimeException("company not found with id :" + id);
+        }
         companyRepository.deleteById(id);
     }
 }
